requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.1
requests==2.31.0
mysql-connector-python==8.3.0
whitenoise==6.6.0
gunicorn==21.2.0 